moved shared rpc logic to single file.
